simplify signup flow
