+++ components/login/LoginPageContent.tsx
@@ -12,14 +12,9 @@ import { setUrlWithoutRerender } from 'lib/utilities/browser';
 import type { ErrorType } from 'lib/utilities/errors';
 import splashImage from 'public/images/artwork/world.png';
 
+import { Container } from './components/LoginLayout';
 import { Login } from './Login';
 import { LoginErrorModal } from './LoginErrorModal';
-
-export const Container = styled(Box)`
-  max-width: 100%;
-  width: 1170px;
-  margin: 0 auto;
-`;
 
 type Props = {
   hideLoginOptions?: boolean;
