+++ lib/public-api/createWorkspaceApi.ts
@@ -7,8 +7,8 @@ import { upsertUserForDiscordId } from 'lib/discord/upsertUserForDiscordId';
 import { upsertUserRolesFromDiscord } from 'lib/discord/upsertUserRolesFromDiscord';
 import { trackUserAction } from 'lib/metrics/mixpanel/trackUserAction';
 import { updateTrackGroupProfile } from 'lib/metrics/mixpanel/updateTrackGroupProfile';
-import type { SpaceCreateInput } from 'lib/spaces/createWorkspace';
-import { createWorkspace } from 'lib/spaces/createWorkspace';
+import type { SpaceCreateInput } from 'lib/spaces/createSpace';
+import { createWorkspace } from 'lib/spaces/createSpace';
 import { getAvailableDomainName } from 'lib/spaces/getAvailableDomainName';
 import { createUserFromWallet } from 'lib/users/createUser';
 import { InvalidInputError } from 'lib/utilities/errors';
