+++ pages/index.tsx
@@ -39,8 +39,8 @@ export default function LoginPage() {
       router.push(returnUrl);
     } else if (spaces.length === 0 && !isSpaceDomain(returnUrl?.replaceAll('/', ''))) {
       // Note that a user logging in will be redirected to /signup, because the 'user' and 'spaces' are loaded async after the wallet address appears.
-      log.info('Redirect user to signup');
-      router.push('/signup');
+      log.info('Redirect user to create a space');
+      router.push('/createSpace');
       // send to signup for users without a workspace unless they are being redirected to an existing workspace
     } else if (returnUrl) {
       log.info('Redirect user to given url');
